ui_services/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
djangogrpcframework==0.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
grpcio==1.34.0
grpcio-tools==1.34.0
idna==2.10
protobuf==3.14.0
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
